Move `options.proto` back to the root package.
