+++ codegen-plugin/src/main/kotlin/io/spine/examples/protodata/hello/plugin/ApplySizeOptionPlugin.kt
@@ -25,7 +25,7 @@
  */
 package io.spine.examples.protodata.hello.plugin
 
-import io.spine.examples.protodata.hello.option.ArrayOfSizeOption
+import io.spine.examples.protodata.hello.ArrayOfSizeOption
 import io.spine.protodata.plugin.Plugin
 import io.spine.protodata.plugin.ViewRepository
 import io.spine.protodata.renderer.Renderer
