+++ codegen-plugin/src/main/kotlin/io/spine/examples/protodata/hello/plugin/SizeOptionView.kt
@@ -28,7 +28,7 @@ package io.spine.examples.protodata.hello.plugin
 import io.spine.core.External
 import io.spine.core.Subscribe
 import io.spine.core.Where
-import io.spine.examples.protodata.hello.option.ArrayOfSizeOption
+import io.spine.examples.protodata.hello.ArrayOfSizeOption
 import io.spine.protobuf.AnyPacker.unpack
 import io.spine.protodata.event.FieldOptionDiscovered
 import io.spine.protodata.plugin.View
